Create ApiWrapperClient to keep code reusable
