+++ core/src/main/kotlin/org/dreamexposure/discal/core/business/api/CamApiWrapper.kt
@@ -1,14 +1,10 @@
 package org.dreamexposure.discal.core.business.api
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import discord4j.common.util.Snowflake
 import okhttp3.HttpUrl.Companion.toHttpUrl
-import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
-import okhttp3.Response
-import okhttp3.executeAsync
 import org.dreamexposure.discal.core.config.Config
 import org.dreamexposure.discal.core.enums.calendar.CalendarHost
 import org.dreamexposure.discal.core.logger.LOGGER
@@ -16,13 +12,12 @@ import org.dreamexposure.discal.core.`object`.new.model.ResponseModel
 import org.dreamexposure.discal.core.`object`.new.model.discal.cam.SecurityValidateV1Request
 import org.dreamexposure.discal.core.`object`.new.model.discal.cam.SecurityValidateV1Response
 import org.dreamexposure.discal.core.`object`.new.model.discal.cam.TokenV1Model
-import org.dreamexposure.discal.core.`object`.rest.ErrorResponse
 import org.dreamexposure.discal.core.utils.GlobalVal.JSON
 import org.springframework.stereotype.Component
 
 @Component
 class CamApiWrapper(
-    private val httpClient: OkHttpClient,
+    private val apiWrapperClient: ApiWrapperClient,
     private val objectMapper: ObjectMapper,
 ) {
     private final val CAM_URL = Config.URL_CAM.getString()
@@ -36,7 +31,7 @@ class CamApiWrapper(
             .header("Content-Type", "application/json")
             .build()
 
-        return makeRequest(request, SecurityValidateV1Response::class.java)
+        return apiWrapperClient.makeRequest(request, SecurityValidateV1Response::class.java)
     }
 
     suspend fun getCalendarToken(credentialId: Int): ResponseModel<TokenV1Model> {
@@ -52,7 +47,7 @@ class CamApiWrapper(
             .url(url)
             .build()
 
-        return makeRequest(request, TokenV1Model::class.java)
+        return apiWrapperClient.makeRequest(request, TokenV1Model::class.java)
     }
 
     suspend fun getCalendarToken(guildId: Snowflake, calNumber: Int, host: CalendarHost): ResponseModel<TokenV1Model> {
@@ -69,38 +64,6 @@ class CamApiWrapper(
             .url(url)
             .build()
 
-        return makeRequest(request, TokenV1Model::class.java)
-    }
-
-    private suspend fun <T> makeRequest(request: Request, valueType: Class<T>): ResponseModel<T> {
-        var response: Response? = null
-
-        try {
-            response = httpClient.newCall(request).executeAsync()
-
-            when (response.code) {
-                200 -> {
-                    val data = objectMapper.readValue(response.body!!.string(), valueType)
-                    response.body?.close()
-                    response.close()
-
-                    return ResponseModel(data)
-                }
-                else -> {
-                    val error = objectMapper.readValue<ErrorResponse>(response.body!!.string())
-                    response.body?.close()
-                    response.close()
-
-                    return ResponseModel(error, response.code)
-                }
-            }
-
-        } catch (ex: Exception) {
-            LOGGER.error("Error making request host:${request.url.host} | uri:${request.url.encodedPath} | code:${response?.code}", ex)
-            throw ex // Rethrow and let implementation decide proper handling for exception
-        } finally {
-            response?.body?.close()
-            response?.close()
-        }
+        return apiWrapperClient.makeRequest(request, TokenV1Model::class.java)
     }
 }
